+++ src/main/java/com/example/social_network01/controller/RepostController.java
@@ -1,9 +1,13 @@
 package com.example.social_network01.controller;
 
+
 import com.example.social_network01.dto.RepostDTO;
+import com.example.social_network01.model.User;
 import com.example.social_network01.service.repost.RepostService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,7 +20,9 @@ public class RepostController {
     private RepostService repostService;
 
     @PostMapping
-    public RepostDTO createRepost(@RequestBody RepostDTO repostDTO) {
+    public RepostDTO createRepost(@AuthenticationPrincipal User currentUser,
+                                  @RequestBody RepostDTO repostDTO) {
+        repostDTO.setUserId(currentUser.getId());
         return repostService.createRepost(repostDTO);
     }
 
@@ -36,8 +42,17 @@ public class RepostController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteRepost(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteRepost(@AuthenticationPrincipal User currentUser, @PathVariable Long id) {
+
+        RepostDTO repost = repostService.getRepostById(id);
+
+        // Проверяем, что пользователь - автор сообщения
+        if (!repost.getUserId().equals(currentUser.getId())) {
+            throw new AccessDeniedException("You can't edit this repost");
+        }
+
         repostService.deleteRepost(id);
         return ResponseEntity.noContent().build();
+
     }
 }
